+++ golang/networks/network_context.go
@@ -11,16 +11,14 @@ import (
 	"github.com/kurtosis-tech/kurtosis-client/golang/services"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
-	"io"
 	"os"
-	"path"
 	"sync"
 )
 
 type PartitionID string
 
 const (
-	// This will alwyas resolve to the default partition ID (regardless of whether such a partition exists in the network,
+	// This will always resolve to the default partition ID (regardless of whether such a partition exists in the network,
 	//  or it was repartitioned away)
 	defaultPartitionId PartitionID = ""
 )
@@ -315,56 +313,5 @@ func (networkCtx *NetworkContext) ExecuteBulkCommands(bulkCommandsJson string) e
 	if _, err := networkCtx.client.ExecuteBulkCommands(context.Background(), args); err != nil {
 		return stacktrace.Propagate(err, "An error occurred executing the following bulk commands: %v", bulkCommandsJson)
 	}
-	return nil
-}
-
-// Docs available at https://docs.kurtosistech.com/kurtosis-libs/lib-documentation
-func (networkCtx *NetworkContext) RegisterStaticFiles(ctx context.Context, staticFileFilepaths map[services.StaticFileID]string) error {
-	// Sanity-check that the filepaths exist
-	for staticFileId, absFilepathOnTestsuite := range staticFileFilepaths {
-		if _, err := os.Stat(absFilepathOnTestsuite); os.IsNotExist(err) {
-			return stacktrace.NewError("Filepath '%v' associated with static file '%v' doesn't exist", absFilepathOnTestsuite, staticFileId)
-		}
-	}
-
-	staticFileIdStrsToRegister := map[string]bool{}
-	for staticFileId := range staticFileFilepaths {
-		staticFileIdStrsToRegister[string(staticFileId)] = true
-	}
-	args := &core_api_bindings.RegisterStaticFilesArgs{
-		StaticFilesToRegister: staticFileIdStrsToRegister,
-	}
-	resp, err := networkCtx.client.RegisterStaticFiles(ctx, args)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred registering static file IDs: %+v", staticFileIdStrsToRegister)
-	}
-
-	// Now that we have destinations where the files should live, copy them there
-	for staticFileIdStr, destRelativeFilepath := range resp.StaticFileRelativeFilepaths {
-		staticFileId := services.StaticFileID(staticFileIdStr)
-		srcAbsFilepathOnTestsuite, found := staticFileFilepaths[staticFileId]
-		if !found {
-			// This should never happen!
-			return stacktrace.NewError("Static file ID '%v' returned by the Kurtosis API should be in the source file map, but wasn't - this is definitely a bug in Kurtosis!")
-		}
-		destAbsFilepathOnTestsuite := path.Join(networkCtx.suiteExVolMountpoint, destRelativeFilepath)
-		if _, err := os.Stat(destAbsFilepathOnTestsuite); err == nil {
-			return stacktrace.NewError("When registering static file '%v', the Kurtosis API returned a path '%v' in the suite execution volume where a file already exists - this is a bug in Kurtosis!", staticFileId, destRelativeFilepath)
-		}
-		destFp, err := os.Create(destAbsFilepathOnTestsuite)
-		if err != nil {
-			return stacktrace.Propagate(err, "An error occurred opening static file '%v' destination file '%v' for writing", staticFileId, destAbsFilepathOnTestsuite)
-		}
-		defer destFp.Close()
-		srcFp, err := os.Open(srcAbsFilepathOnTestsuite)
-		if err != nil {
-			return stacktrace.Propagate(err, "An error occurred opening static file '%v' source file '%v' for reading", staticFileId, srcAbsFilepathOnTestsuite)
-		}
-		defer srcFp.Close()
-		if _, err := io.Copy(destFp, srcFp); err != nil {
-			return stacktrace.Propagate(err, "An error occurred copying all the bytes from static file '%v' source filepath '%v' to destination filepath '%v'", staticFileId, srcAbsFilepathOnTestsuite, destAbsFilepathOnTestsuite)
-		}
-	}
-
 	return nil
 }